requirements.txt
livekit-agents[codecs,images]==1.1.3
livekit-plugins-openai==1.1.3
livekit-plugins-silero==1.1.3
python-dotenv>=1.0
